fix:  Fix PlatformIO builds
Fixed PlatformIO builds to work with new folder structure.  Removed submodule python script in favor of just ignoring the directory.  This will allow desktop and IoT to better co-exist in the same source tree. Remove [common] section of test/platformio.ini as it did not seem to actually be respected.