fix: Arduino HTTP Request String
This commit adds a 'toHttpStr' method to properly prepend an HTTP request string with "https://".

Arduino's HTTPClient requires a single-line URL request string begin with "http://".

Absence of "http://" will lead to 'connection refused' errors.

This is a client-side issue and is only a consideration on IoT, so this should not be resolved anywhere else except this header.