chore: Update build to be out of source
Updated build to be out of source to clean up the workspace
and added build scripts to enforce this.

Re-enabled  cmake_example builds in CI.

Removed cmake_example .gitmodules file.  This historically caused
git to do funny things having a nested .gitmodules in the source
tree.  Commented out the submodule clone in the CMakeList.txt and
setup the build script to mimic the submodule clone by copying the
cpp-crypto source into the correct directory.