ci: split coverage workflows
As in the C++ Crypto repo, CodeClimate Coverage attempts to use the CC_TEST_REPORTER_ID of the PR base for reporting on the upstream parent and fails.

This PR splits workflows so that CodeClimate Coverage onlys runs on Pushes to master/devleop.

CodeCov will continue to run on PR's and Pushes.